Refine needsWriteBarrier to consider HV types
Summary:
We assert that whenever `NeedsBarriers` is set to false,
`needsWriteBarrier` must be false. However, the current implementation
of `needsWriteBarrier` in Hades is fairly rudimentary, it returns true
for all locations in the OG.

In SH, we would like to elide write barriers when the before and after
type of a `HermesValue` is known to not be a pointer or symbol type.
Allow `needsWriteBarrier` to consider this distinction.

Differential Revision: D42589793

fbshipit-source-id: 2849d17ec8717636fb9408ac07eefdcac10667ed