Remove register stack size override in hermes.cpp
Summary:
We are currently overriding any specified register stack size with a
size that is fairly low. This was added to allow the register stack to
be placed on a thread stack when StackRuntime is enabled, however, it
seems like that functionality was removed some time ago, so this limit
no longer serves a purpose.

Remove it so that sizes set through RuntimeConfig are actually
honoured. Set the default RuntimeConfig value to be roughly what the
old limit was (no need to suddenly increase the stack size if we
haven't had problems yet). Keeping the limit low helps us retain the
ability to place the register stack in StackRuntime again in the future.

Reviewed By: mhorowitz

Differential Revision: D42610775

fbshipit-source-id: 94db4e3eb1d6365749a6db8adfe73a5d298f127c