Make SimplifyCFG a ModulePass and delete unused functions
Summary:
Our IR has an invariant that in order to load from a frame variable,
that variable must have stores to it. However, this is complicated by
unreachable functions, since optimisations that render them unreachable
may still leave them in the IR. For example, this becomes a problem in
SimplifyCFG in the following case:
```
function foo() {
  if (false) {
    let x = 10;
    let inner = function() { return x; }
  }
  return 0;
}
```
SimplifyCFG can detect that the code inside the `if` will never
execute, and so delete the store to `x`. And while it also renders the
inner function unreachable, it doesn't actually delete it from the
`Module`. This can cause problems in other passes (e.g. type inference)
which rely on loads from variables having at least one corresponding
store.

To fix this, delete all unused functions at the end of SimplifyCFG.

Reviewed By: tmikov

Differential Revision: D42831464

fbshipit-source-id: ee61edef9a6e6b3adc24fe8326f0a21b414657a8