Refine needsWriteBarrier to consider HV types (#895)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/895

We assert that whenever `NeedsBarriers` is set to false,
`needsWriteBarrier` must be false. However, the current implementation
of `needsWriteBarrier` in Hades is fairly rudimentary, it returns true
for all locations in the OG.

In SH, we would like to elide write barriers when the before and after
type of a `HermesValue` is known to not be a pointer or symbol type.
Allow `needsWriteBarrier` to consider this distinction.

Reviewed By: tmikov

Differential Revision: D42589793

fbshipit-source-id: 5d17522bbb4078740a8e4a68c140bd1fba8be3a5