Implement a simple stack promotion pass
Summary:
The existing StackPromotion pass is difficult to understand. It also
interacts poorly with other passes such as TDZDedup, because its
complexity results in the addition of invariants and assumptions which
are hard to reason about.

Much of the complexity stems from StackPromotion's desire to partially
promote variables if they are known to not be captured within a given
block.

Forwarding known non-literal expressions is also fraught with
complexity, and trying to do it correctly in `StackPromotion` is too
risky. We have to consider the function where the load/store occur,
recursive invocations, dominance relationships between the load/store,
and so on.

To remedy this, add a SimpleStackPromotion pass that splits promotion
into three steps:
1. Replace all variables that only ever contain a single literal value
with that literal value.
2. If a variable is only ever stored to from the function that owns it,
create a copy of the variable on the stack in the owning function. This
allows loads from the frame in the owning function to be replaced with
loads from the stack.
3. If a variable is never loaded from, delete all stores to it.

(2) on its own is intended to replace the current "const" variable
promotion, where if a variable has a single known non-literal
expression, that value may be forwarded to a load. The process of
actually forwarding the expression is then left to Mem2Reg. This can
actually result in better code than the current `StackPromotion`, since
variables with multiple stores can also be forwarded (see an example in
the next diff).

(3) is also an improvement over the current `StackPromotion`, which
does not eliminate store-only variables.

(2) and (3) combined have the net effect of what we typically consider
"stack promotion". For variables that are never captured, (2) will
eliminate all loads from them, and then (3) will eliminate all stores
(since there are no remaining loads).

Reviewed By: tmikov

Differential Revision: D42360753

fbshipit-source-id: 4eb66f31565e3c5b05a4f7e717b4b4283d7a8f12