Replace SemanticValidator with SemanticResolver
Summary:
`SemanticValidator` performed some AST validation, but did not resolve
identifiers. `SemanticResolver` performs the same validations, but also
resolves all identifiers, following ES6+ scoping rules.

## Administrative Changes

Remove the ESTree decorator provisions for running both validators and
exclude the old validator from the build (which requires various small
changes all over the source tree). However, do not delete the old
validator files yet, preserve them without compiling them. Specifically:
- move `include/hermes/AST/SemValidate.h` to `lib/AST/`.
- exclude `lib/AST/SemValidate.cpp` and `lib/AST/SemanticValidator.cpp`
  from the build.

## Non-recursive Compilation

Since all identifiers have been resolved by the resolver, we no longer
need to maintain state for all surrounding functions. Each function has
enough data to be compiled on its own.

So, instead of compiling functions recursively, as we encounter them, we
instead just create an empty Function value and enqueue a callback in a
*compilation queue*. When callback is eventually invoked, it compiles
the function and populates its body.

As a result, functions are output in a different order and their
locality can be affected. However locality shouldn't be an artifact of
arbitrary decisions made by IRGen anyway.

## Variable Resolution

The most substantive change is that IRGen no longer needs to maintain
scope-based lookup tables, instead using the resolution information
generated by `SemanticResolver`. Specifically, every resolved
`ESTree::IdentifierNode` contains a pointer to the corresponding
`sema::Decl`.

Additionally, `sema::Decl` contains an extra field
`void *customData`, which is used by IRGen to associate an IR `Valuei *`
with the `Decl` (either `Variable *` or `AllocStackInst *`). The latter
eliminates the need for any additional lookup tables.

Upon entering a new lexical scope, IRGen iterates the list of
declarations stored in the `ScopeDecorationBase` decoration and creates
the corresponding IR values.

This produces correct ES6 lexical scope resolution, but not the
correct capturing semantics in loops. The latter requires further
non-trivial changes to the IR and will be implemented later.

## "Scoped" For-loop Semantics

For loops (for/for-of/for-in) with declarations in the init expression
have complex semantics in ES6+. Specifically in `for`, every iteration
creates a new lexical scope, where the test condition is evaluated in
the current iteration's scope, but the update expression is evaluated in
the next iteration's scope, and the values have to be propagated from
one scope to the next.

The complexity has forced us to compile "scoped for-loop"
 `for(declaration; ...;...)` very differently from the normal form. The
"scoped" version produces rather unnatural and hard to optimize code,
which we hope is correct. It may be possible to possible to produce much
better code, if we can detect whether the for-loop update expression
captures any variables. This should happen in followup diffs.

The correct semantics of scoped `for-of` and `for-in` **have not been
implemented yet**.

## Other Changes

Remove support for lazy compilation and compilation in a scope. It will
be implemented later from first principles.

Remove `Variable::DeclKind` and use `Variable::obeysTDZ_` and
`Variable::isConst_` instead (the latter may not actually even be
needed).

Global property declarations are now handled explicitly in the IR. Old
code:
```
function global() : undefined
frame = [], globals = [a, b]
%BB0:
  %0 = ReturnInst undefined : undefined
```
New code:
```
function global() : undefined
frame = []
%BB0:
  %0 = DeclareGlobalVarInst "a" : string
  %1 = DeclareGlobalVarInst "b" : string
  %2 = ReturnInst undefined : undefined
```

## Important TODO

TODO in follow up diffs:
- implement correct semantics in for-in and for-of loops.
- implement true scopes for captures inside loops.
- implement for-loop update capturing heuristic.
- run test262 on this commit.

Differential Revision: D42247692

fbshipit-source-id: e4d79aa32e84e00be611127e9fb45769d38c661a