Sema: improve handling of "arguments"
Summary:
Due to limitations in IRGen, we can't support assignment to "arguments",
even when it is allowed. So, add code to reject it.

Additionally, we cannot support the obscure case `var arguments`, which
is just an alias for the existing arguments, because it could have an
initializer, which goes back to the previous point.

Instead we treat `var arguments` as a new variable, which naturally
allows modification.

This is non-compliant, but Hermes has the same behavior already.

Lastly, fix the behavior of arrow functions, when the captured
"arguments" is just a normal declaration instead of *the* arguments
object.

To do this properly, always declare the arguments object in a function,
unless it is an arrow, has a parameter named "arguments", or a
declaration named "arguments". This allows "arguments" to be resolved as
any other identifier without special cases.

There is an additional complication in the following code:
```
function foo(a=arguments[1]) {
  let arguments = 10;
  print(arguments);
}
```
The "arguments" in the initializer expression is the arguments object,
while the "arguments" in the `print()` is the `let` declaration. So, if
the right conditions are met, we declare the arguments object during
parameter resolution, then we *unbind* it from the binding table. After
the body declarations have been processed, we can decide whether to
re-bind it again, or leave it unbound, since there is a declaration in
the body like in the example above.

Add a few tests.

Add a new document explaining deliberate decisions to be incompatible
with the spec.

Reviewed By: avp

Differential Revision: D42952096

fbshipit-source-id: cfcfe7e60daeec7710eeb818d3d4c36e99255a04