Differentiate between declaration decl and expression decl
Summary:
In one case that we have identified so far, the variable declared by a
declaration and the variable that is written to by the initializer are
different:

```
  try {
  } catch (e) {
    var e = 10;
    print(e);
  }
```

The `var` declares the variable `e`, which is hoisted to the
function/global scope. However the `e = 10` is resolved not to the
hoisted variable, but to the catch parameter `e`. The `e` in `print(e)`
is also resolved to the catch parameter.

So we have an AST node - the identifier in the variable declarator -
which is resolved to two different variables, or "decls" in Sema-speak,
depending on the context. For lack of better ideas, we call the
*declaration decl* and the *expression decl*.

As noted, this is an extremely rare occurrence, so it is not worth it to
add two pointers to every AST node. Instead, we add flags indicating
whether either of the *declaration decl* or *expression decl* are set
and whether they are set to the same value. If they are both set, but to
a different value, we store the value of the *declaration decl* in a
side table.

Modify Sema AST printer to print this data. Update IRGen
correspondingly.

Reviewed By: avp

Differential Revision: D42912828

fbshipit-source-id: 4f5538274b1cba079499d7a8264a23ff23a78049