Fix ScopedFunctionPromoter
Summary:
`ScopedFunctionPromoter` wasn't querying the correct AST nodes for their
scope declarations as it was running, because it was calling
`visitESTreeChildren` on the function nodes themselves.
It's supposed to call on the function bodies so they are associated
with the function scope.
Fix this by using `getBlockStatement`.

This won't work on `ProgramNode`, which doesn't have a `BlockStatement`.
The function wasn't getting exercised anyways because there was no call
from `SemanticResolver` in `ProgramNode`'s visitor, so add that too.

So now we get a global declaration for `foo` on the following code:
```
foo;
{
  function foo() {}
}
```

Some observations:
1. The DeclCollector sees `function foo` and adds it to the list of
`scopedFuncDecls_`.
2. The `SemanticResolver` does _not_ set the `Kind` of the actual
declaration to `ScopedFunction` but rather `GlobalProperty`.
3. `ScopedFunctionPromoter` performs promotion because it iterates
`scopedFuncDecls_`.
4. An error is reported based on the `Kind`, which does conflict with
`let`.

To avoid this, we can check if the function declaration is really in
the root scope of the global function following promotion.
If it is, then it's a `GlobalProperty`, otherwise it remains as a
block-scoped function in a child scope within the global function.

Reviewed By: tmikov

Differential Revision: D42434677

fbshipit-source-id: 31c0aeea2677350f6ec5273cf218be24552ae7bd