Add Function attributes
Summary:
To allow encoding a call graph in the IR and maybe more things in
the future, add an `Attributes` to each Function.

The `Attributes` for now only has two flags, one of which is actually
going to be populated in the near future.

* `allCallsitesKnown` can be used for inlining (e.g.)
  and indicates whether all callsites of the function are certain
  to have their `target` operands set to the function.
* `pure` is included for illustrative purposes and can be used in the
  future for some currently unknown notion of "purity".

This information should also be added to the Dumper at some point,
but I'm holding off on that to see how the Dumper changes with
all the other IR changes to avoid too many conflicting changes.

Differential Revision: D42318251

fbshipit-source-id: 3bef69c74cd0e608a5d6e786d30680cc3ed26636