Use call target information when inlining
Summary:
Use the IR's call graph to create an ordering to visit functions
in the inlining pass.

We start at any functions we can find which we don't think have any
known callsites, and seed the search with that first.
Using those as a starting point,
we use the postorder visitor from each function to try and ensure that
functions are visited before the callers for those functions,
which will be conducive to inlining.
This enables inlining sibling functions, as demonstrated in the test.

Then, we reorganize the pass itself to use the call graph and the
derived ordering on `Function`s, delete the code for figuring out
callsites ad-hoc, and preserve semantics by checking for `CallInst`
instead of some other instruction, and explicitly checking for recursion
because the call graph itself will register the callee (it is correct
to register the target operand, we just can't inline it, and technically
it's dead code anyway).

Reviewed By: neildhar

Differential Revision: D40962637

fbshipit-source-id: 786b7e4fd8debfa696a8caf82b269e6ba91d54de