Define parsing of VM command line options in libhermesvm
Summary:
`_sh_init(int argc, char **argv)` is a convenient way to initialize the
VM in CLI tools, but it had serious problems:
- Runtime creation was not thread safe, since it relied on a dense map
of runtimes.
- There was no way to initialize a SHRuntime without parsing command
line options, since the default config was constructed from
`llvh::cl::opt` instances.
- Calling `_sh_init()` more than once would "merge" the previous state
of the command line parser with the new one.
- Command line options as global variables add startup time, relocations
and size on mobile, and made it hard to control when and where they were
registered with the CLI parser.
- Command line parsing itself was not thread safe.

Address these problems methodically:
1. Make SHRuntime creation and destruction thread safe.
2. Expose a `_sh_init()` overload taking a `vm::RuntimeConfig`, no
parsing of CLI.
2. Define the command line flags used for VM configuration in structs,
that can be conveniently and explicitly instantiated when and where
needed. The structs are split in two: `VMOnlyRuntimeFlags` containing,
as its name suggests, only flags relevant to the VM, and `RuntimeFlags`,
containing those plus flags shared between the compiler and VM. Users of
these structs can instantiate either, depending on need.
3. Serialize command line parsing with a mutex, making it thread-safe,
and reset the global CLI parser state before and after using it.
4. Define a `_sh_init(argc, argv)` overload returning an error instead
of terminating.

Reviewed By: neildhar

Differential Revision: D43108112

fbshipit-source-id: f85402de91f525a3a1936083117995b495e72709