Wire in options and APIs for type checking
Summary:
Extend SemResolve to support type checking by passing an optional
FlowContext as a parameter. It doesn't do the actual type check yet, but
the wiring is there.

New API to dump the context with the optional types. The types will be
generated by the type checker in a later diff, but the wiring is there.

Finally, modify the `shermes` tool to invoke the type checking APIs. A
new `--typed` option enables type checking and auto-enables
`--parse-flow` unless `--parse-ts` is enabled. It also auto-enables
`-fstatic-builtins` unless it is explicitly disabled.

Reviewed By: avp

Differential Revision: D42492190

fbshipit-source-id: 1c2eaedf10d2b60610fef8353f51aa690d5f6d0f