Add typed store inline methods
Summary:
Add inline methods for storing some basic types, when we know that the
value being stored and the destination have that type.

Update code generation to use the new inline methods. This is written
with the expectation that the stored value operand to the PrStore has
the same type as the property being stored to.

Reviewed By: tmikov

Differential Revision: D42928970

fbshipit-source-id: 4833ad1f6e03c121712fa8e3323ceb9a12087b89