Remove case-sensitive paths plugin
Summary:
Recent versions of docusaurus seem to mix case for some type names, causing a conflict with the "case sensitive paths" plugin. Removing this plugin resolves the issue (and restores the ability to view the site locally), but I'm not sure why it was added in the first place (it seems to date back to the prototype of the webpage).

The references to the mixed-case type seem to be entirely within docusaurus (i.e. not the Hermes site itself).

Reviewed By: jpporto

Differential Revision: D43330318

fbshipit-source-id: d19dab1cd5fc1c1e5bb86f35f84290210c6a8e5c