Precompute if a string is ASCII or not
Summary:
`StringPrimitive::create` internally needs to check if a string is
composed of all ASCII characters. It does this by iterating over all
the characters and checking to see if they are within bounds. Given
that we are iterating over each character during parsing, we can keep
track of this information during the loop and pass it into `
StringPrimitive::create`.
This results in ~4% perf improvement for JSON
parsing.

Reviewed By: jpporto

Differential Revision: D40740681

fbshipit-source-id: cc16082d7372b853e9272686d83c78a817c236dc