Implement constructor super() call
Summary:
`super()` must be called in the first line of the constructor of any
derived class. Ordinarily, we'd allow it to be called anywhere prior to
the first use of `this`, but this extreme restriction greatly simplifies
verification.

Add checks to ensure that `super()` is only called once in the
constructor by storing the super call in the FunctionContext in the
SemanticResolver.

In IRGen, we store the current class being generated so we can look up
its superclass.
Skip generating implicit constructors for derived functions for now,
require users to specify them manually until we implement them.

Reviewed By: tmikov

Differential Revision: D43505402

fbshipit-source-id: 5362842ba2ef579de6cb08a43fe9a418a6fed3f3