Implement inheritance for methods
Summary:
In the typechecker, we have to visit the superClass node prior to the
actual `parseClass` call so we have the superClass AST node resolved.
When declaring methods in the derived class, first check and see if
they are overriding and place them in the existing position.
Otherwise, add a new position with the method.

The lookup table contains all the self and inherited fields,
which are copied in during initialization.
This enables quick lookup during typechecking,
and use of the `layoutSlotIR` means we can just iterate the lookup table
and fill in the slots on the prop map as we go.

The HomeObject creation must also set its parent to the superClass's
HomeObject so that eventually `super.foo()` can work.

No constructor chaining with `super` in this diff,
so no field inheritance in this diff.

Reviewed By: tmikov

Differential Revision: D43505399

fbshipit-source-id: 95608a687f21157eae13b0cc7df0ac34f2486337