Remove instanceof fast path
Summary:
`instanceof`'s fast path is incorrect, it elides the lookup of
`Symbol.hasInstance` if the RHS is a function, based on the reasoning
that `Function.prototype[Symbol.hasInstance]` cannot be modified.
However this is incorrect since you could still define
`Symbol.hasInstance` on the function itself, or change the prototype of
the function.

Reviewed By: avp

Differential Revision: D43642297

fbshipit-source-id: 07ea77ee3459c538013f024f8d2e16bfb85bd682