Correct type inference for ThrowIfEmpty
Summary:
With a statically typed system, we need ThrowIfEmpty to *filter out* the
empty from the input, so the potentially statically typed
instructions after it can consume it.

In the following case:

        let x: number;
        function foo() { return x + 1; }

the actual type of `x` is `number | empty`, because of TDZ. However the
statically typed "add" instruction requires its two operands to be the
same number type. So, `ThrowIfEmpty` "filters out" the Empty type from
its input, returning the input type minus the Empty type.

Implement this in the constructor of `ThrowIfEmpty` and in
TypeInference.

A subtle problem arises in the case where the compiler is able to prove
that `ThrowIfEmpty` will always throw, because its operand type is just
Empty, as can be seen here:

        x + 3.14;
        let x : number = 10;

A hypothetical IR will look something like:

             StoreFrameInst empty: empty, [x]: any
        %3 = LoadFrameInst (:empty) [x]: empty
        %4 = ThrowIfEmptyInst (:notype) %3: empty
        %5 = fadd (:number) $4: notype, 3.14: number

We know that `x` has type Empty, then `ThrowIfEmpty` has filtered it out,
correctly producing a NoType as its result. While that is theoretically
correct in some sense, it violates the invariant of the `fadd`
instruction, which requires both of its operands to be type number.

The easiest solution to this problem it to just pretend that
`ThrowIfEmpty` still returns number. This is what this diff does, with
some gymnastics.

Reviewed By: neildhar

Differential Revision: D43808595

fbshipit-source-id: 88bfee6539246f54368eed727649349147c39879