Rename hermes-translate to flow-api-translator
Summary: Rename `hermes-translate` package to `flow-api-translator` to make the name clearer what it does/ does not do. For example this tool is not a general purpose Flow to TS translator. It purposely only focuses on translating the external API surface, relying on Flow to continue checking internal logic, as this is much more simpler and produces better results.

Reviewed By: noahlemen

Differential Revision: D43854719

fbshipit-source-id: 7b66f575994ea984232001e461f05d422d517be2