Modify Inlining to handle functions with "inline" directive
Summary:
Handle multiple callsites and allow only functions marked as "inline"
to be inlined into multiple callsites (due to the obvious code size
impact).

Most of the diff is from the reformat due to indentation.
The real change is to how `callsites` is populated and how it's
iterated instead of the first element being the only one operated on.

Reviewed By: neildhar

Differential Revision: D43845609

fbshipit-source-id: ef9217cf83b9f079df6763f45d379d0441811a68