Add "inline" directive
Summary:
Add detection and propagation of a new "inline" directive.
This will act as a compiler hint to prioritize inlining over code size
for the function it is applied to.

It won't always be possible to inline the function marked "inline",
but it will be attempted more aggressively (even if the function has
multiple callsites, e.g.).

Reviewed By: tmikov

Differential Revision: D43845608

fbshipit-source-id: 2845260f1699958d3abd57a034523ab95420ac7e