Add a run of Mem2Reg closer to SimpleStackPromotion
Summary:
We currently re-run Mem2Reg very late in the pipeline, which means that
new stack variables introduced by StackPromotion in the middle of the
pipeline are harder to optimise. For instance, if inlining eliminates a
frame variable, it only gets converted to an SSA value late in the
pipeline, which prevents the new stack location from being optimised
out.

Reviewed By: avp

Differential Revision: D43895704

fbshipit-source-id: a0da834acae6248d1dbb88b5eeeb6d9437c7ed2f