Process key strings efficiently
Summary:
If we are parsing a string that will be used as a key in an object,
then we know the string hash must be computed eventually, since it will
need to be turned into a SymbolID. However, if we are parsing the
string as a simple string value, then we can make the optimization of
side stepping the hash computation and lookup, and construct a new
StringPrimitive directly. ~7.2% perf improvement.

Reviewed By: tmikov

Differential Revision: D43486478

fbshipit-source-id: 696d7115938594c0f59036a493e232732fa9052e