Inline unreachable basic blocks
Summary:
Inlining skipped unreachable basic blocks because there's not much value
in inlining them.

However, doing so could lead to errors if we have a PhiInst
being inlined where one of the entries comes from an unreachable block.

To handle this we either have to perform the same analysis that
SimplifyCFG does during inlining to avoid breaking invariants,
or simply inline everything and let the later passes clean it up.

Reviewed By: tmikov

Differential Revision: D43987572

fbshipit-source-id: 3ccb4e96547153378051a6cbde242f65fca66da5