LoadStackInst uses the type of its operand
Summary:
TypeInference already sets the type of LoadStackInst to the type of its
operand, but we need it earlier for correctness.

Additionally, SimpleStackPromotion needs to preserve the type when
copying a variable from the frame to the stack. This type now gets
propagated into all loads due to the above change.

This becomes more important up the stack where we introduce empty-type
correctness.

Reviewed By: avp

Differential Revision: D43912459

fbshipit-source-id: 95e2ddcb505eff3a62a2cbbd6f3506c9b5793bf1