Remove unreachable blocks in Inlining
Summary:
Inlining all the unreachable blocks isn't trivial.
They can depend on each other and we can't easily have an entrypoint
to run DFS from.

To avoid running into problems with Phis with unreachable BB operands,
prune the unreachable BB operands by querying the DFS ordering.
If a Phi operand is not in the ordering, it must be unreachable.
It won't be inlined, so ignore it when translating the Phi.

To facilitate this, we turn the `SmallVector` returned by orderDFS
into a `SmallSetVector` for easy membership queries.

Reviewed By: neildhar

Differential Revision: D44039579

fbshipit-source-id: 3e215dad5966b891e0ee232f090de731c66985b9