Remove accidental hermes-eslint dependency
Summary: The `hermes-parser` package imported the visitor keys from the `hermes-eslint` package but did not list `hermes-eslint` as a dependency. This caused module not found errors if you only installed the `hermes-parser`. This change moves the `ESLintVisitorKeys` into the `hermes-parser` package so `hemes-eslint` does not need to be a dependency (and this saves us from adding a circular dependency).

Reviewed By: noahlemen

Differential Revision: D44043710

fbshipit-source-id: ad05550c46b20db888040f8d77a3c12f979bbff4