Avoid recompiling implicit class constructors
Summary:
Implicit class constructors created in `genClassDeclaration` weren't
checking for duplication in the way that all functions ordinarily do
in `genFunctionExpression`.

This led to duplication of class constructor in `finally`,
but only when it was implicit.

Store the implicit class constructor off the `ClassDeclarationNode`
in the `compiledEntities_` map to take care of this.

Reviewed By: tmikov

Differential Revision: D44036958

fbshipit-source-id: 08c14cab0dce9c19ea4a26c4fc33119a96ccbbc6