fix default export handling for globals
Summary: because a GenericTypeAnnotation within a DeclareExportDeclaration expects its reference Identifier to be defined in scope, it currently is unable to handle globals properly, throwing a TranslationError when a global like `$FlowFixMe` or `React$AbstractComponent` appears there. this diff solves this problem by allowing undefined Identifiers to fall through to the export handling with a spoof `$$EXPORT_DEFAULT_DECLARATION$$` VariableDeclaration to allow the type to be exported properly.

Reviewed By: pieterv

Differential Revision: D44109602

fbshipit-source-id: 4581732c62c0bfdd43c3516fcdcc810a99cdbb11