Remove single-create assumption in FunctionScopeAnalysis
Summary:
We now reuse the `Function *` for multiple Create instructions.
Remove the assert that there's a 1:1 relationship and use the first one.

The assert was failing when `HBCResolveEnvironmentInst` was generated
during SH C generation and it called `getScopeDepth`.

All users of `F` that are `BaseCreateLexicalChildInst` will be at the
same scope depth (before block scoping, they're in the same function).
Assert that.

Reviewed By: neildhar

Differential Revision: D44233786

fbshipit-source-id: f520d2e2fb227bb15402bdb56f142dca0964dd5b