Explicitly allocate call registers to the stack
Summary:
Introduce a new register class `RegStack`, representing entries in the
register stack.

Simplify SHRegisterAllocator to track how many register stack entries
are needed for calls.

Change sh::LowerCalls to move call arguments directly into RegStack
registers. Also, extend it to explicitly set the callee and `new.target`
register stack entries (instead of hardcoding them in isel), as well as
to invalidate them appropriately.

Invalidating a register is not fully supported by the IR or the lowered
IR, since we have no concept of "poison" of a "kill" instruction. In
this case, during lowering, we simply use an *implicit mov* of `empty`
into the "killed" register.

Update SH.cpp for these changes. It no longer has to hardcode writes to
the register stack.

Finally, address a long standing benign bug, which became active as a
result of this change. Previously, SH wasn't allocating the *scratch"
register stack entry, which is used primarily by bound calls. So, a
bound call would simply overwrite the offset 1 in the caller frame. It
was benign, since the register stack entries in the caller frame were
unused.

Address this by allocating the two additional stack entries in the frame
and shifting stack indexes correspondingly.

Reviewed By: avp

Differential Revision: D44206516

fbshipit-source-id: 082ca7a6778193e7d2d146c4212df909463b3ba2