Fix running widgets_with_fakemalloc from any directory on macOS
Summary:
On macOS, the dynamic library itself has an "install name" installed at
build time. That name gets used by default when linking against the
library. So because we create the library at `build/libfakemalloc.so`,
that becomes the install name, and ends up being used when we try to
resolve the library.

Reviewed By: tmikov

Differential Revision: D44358247

fbshipit-source-id: 6e70717149cc5d8357c4de3fab179585df180b55