TDZDedup: enforce EmptyType correctness
Summary:
EmptyType should not be propagating to instructions that are not
equipped to handle it (which is most instructions). TDZDedup optimizes
out `ThrowIfEmpty` checks, causing the types of TDZ variables, which
include EmptyType, to reach normal instructions.

Emit `UnionNarrowTrusted` in place of the eliminated `ThrowIfEmpty`.

This fixes the previously disabled test:
- test/Optimizer/es6/tdz-dedup.js

Reviewed By: neildhar

Differential Revision: D44392669

fbshipit-source-id: fa2f9c0c59ee759fc8eb14d9770ccbfd27ce3ab7