IR: add LIRDeadValueInst
Summary:
`LIRDeadValueInst` generates a value of any type, determined by its
result type, where the value will never actually be used at runtime. Its
purpose is only to satisfy the type constraints of downstream
instructions that are generated but will never actually execute.

Yes, in theory we should have eliminated all instructions that we know
won't execute, but we aren't currently, and it feels fragile to have
passes rely on 100% correct DCE. The effort of adding this instruction
is low and we can remove it if we decide we don't need it.

It will be emitted by the next diff.

It has the `LIR` prefix because for now it will only be emitted during
lowering. If that ever changes, we can rename it.

Reviewed By: neildhar

Differential Revision: D44406201

fbshipit-source-id: 9531f2e7bb7295027390a6469a591b3ca2ae7d86