IR: add acceptsEmptyType instruction attribute
Summary:
Annotate instruction with an acceptsEmptyType attribute, indicating
whether the instruction can handle operands of type EmptyType. Most
instructions do not; only few like LoadFrameInst or ThrowIfEmptyInst do.

The attribute allows us to enforce this IR invariant in the verifier.

The attribute itself is implemented as a method on `Instruction` and an
implementation method `acceptsEmptyTypeImpl()` that has to be overridden
only in instructions that must return true.

Reviewed By: neildhar

Differential Revision: D44392663

fbshipit-source-id: 8ca37690c1fb4058c3f05eeb95d5a6c0bccf936e