IR: define UnionNarrowTrustedInst
Summary:
`UnionNarrowTrustedInst`, as it name suggests, "narrows" unions by
removing some of their arms, when the compiler has proven that a value
has a narrower type. This will be used for TDZ elimination, where a
variable type includes "empty", but we know it isn't.

Add support in TypeInference and InstSimplify.

This instruction is similar to ThrowIfEmpty, in that it also needs to
record its initial type.

Reviewed By: neildhar

Differential Revision: D44392665

fbshipit-source-id: 1046f9c6a752d1f2195716706d04e958430d058b