Add isStrict argument to DirectEval
Summary:
Direct calls to eval should inherit the strictness of the function they are defined in. We already have that information available in `ESTreeIRGen-expr.cpp`, so this diff is simply a matter of plumbing that information all the way through to the runtime.

Introduce a new argument to the bytecode operation `DirectEval`, and use this argument in `Interpreter::caseDirectEval`.

Reviewed By: tmikov

Differential Revision: D41860786

fbshipit-source-id: 0faa33e520ccc86d58e280e8242352782583ff28