Improve 'for in' property de-duping
Summary: When collecting all properties in a prototype chain, we must ensure they are de-duplicated. This is because a property in a child object may share the same name as a parent. This diff changes this de-duplication to use llvh::SmallSets, rather than just iterating over the entire array to check for duplicates.

Reviewed By: jpporto

Differential Revision: D43643606

fbshipit-source-id: 29c860617c7ab16741f25e1a87f73b8b4ecdaf27