Mitigate RCEs caused by conversion to double
Summary:
# Context

Hermes method to convert raw numeric values to NaN-boxed Hermes values (`HermesValue::encodeDouble` and `HermesValue::encodeNumber`) are unsafe by default. These functions allow potential NaN values to be encoded in an HermesValue and, when provided an untrusted input, can result in HermesValues encoding objects with arbitrary addresses.
This unsafe behaviour has been the root cause of security incidents in the past.

# Mitigation

This diffs, converges the current encoding API to a safer and simpler one.

The old API **(unsafe functions in bold)**:
- `HermesValue::encodeUntrustedDoubleValue`
- **`HermesValue::encodeDoubleValue`**
- `HermesValue::encodeUntrustedNumberValue`
- **`HermesValue::encodeNumberValue` + overloads**

Is now composed of only two functions:
- `HermesValue::encodeUntrustedDoubleValue`
- **`HermesValue::encodeTrustedDoubleValue`**

The new naming makes unsafe function explicit.
In addition, we:
- migrated most locations to the safe encoding
- applied the unsafe functions to performance-critical parts of the code to reduce

# Performance

NOTE: these runs are old, new ones are running and should be slightly better due to changes on how `encodeUntrustedNumberValue(<integer>)` is implemented.

MobileLab benchmarks show a negligible increase of ~0.06% instructions count (the most stable metrics):

https://www.internalfb.com/mlhub/pipelines/runs/fblearner/413651872
https://www.internalfb.com/mlhub/pipelines/runs/fblearner/413651868

Worth noticing that this is a sensible improvement down from a ~1% increase that would be produced by not having any unsafe function:

https://www.internalfb.com/mlhub/pipelines/runs/fblearner/412793404

The current approach is a good compromise between safety and performance.

Reviewed By: jpporto

Differential Revision: D43363964

fbshipit-source-id: 26221d0b4a471fdde9846181a12c05bd3827894f