IRGen: remove double-init assert
Summary:
The assert mistakenly required variables to be initialized just ince. It
is possible to initialize a variable twice, when it is shadowed by a
parameter or a catch variable. Examples in the newly added tests.

One of the new tests (shadowed-function.js) still fails after this
change, but for a different reason, so it is marked as XFAIL for now.

Reviewed By: neildhar

Differential Revision: D44504976

fbshipit-source-id: fc840283eeebc861ffb14a6265f5eda958049387