IRGen: initialize ScopedFunction declarations to undefined
Summary:
IRGen wasn't initializing ScopedFunction to undefined, since I had
mistakenly thought that they would be initialized as hoisted. This isn't
true when a function is declared in a scope. Its declaration is visible
throughout, but the hoisting only happens in the scope:

    function foo() {
      init = f; // f should be initialized to undefined
      {
        function f() {}
      }
    }

The fix in this diff is to simply always initialize function
declarations to undefined first. Load/store frame optimization should be
able to eliminate the extra stores.

I don't particularly like this solution and it is not really compatible
with typed code, since a typed hoisted function declaration should not
be initialized to undefined at all. But it is a correctness issue and
the fix is simple, so until a better solution is found, this is fine.

Add a new test.

Reviewed By: neildhar

Differential Revision: D44514721

fbshipit-source-id: 390dc8f3d196275582564d83386b6ead5f4b2a25