Ensure formal parameters prevent scoped function promotion
Summary:
The code:
```
function foo(f) {
  {
    function f() {}
  }
  print(typeof f);
}
```
should print `undefined`, because the `function f() {}` shouldn't be
promoted.
The spec notes that "parameterNames does not contain F" is one of the
conditions necessary to allow function promotion, but because the
function params aren't included in DeclCollector they aren't added to
the binding table for detecting collisions in ScopedFunctionPromoter.

To fix this, we do two things:
1. Add all parameters for the function to the ScopedFunctionPromoter
binding table to prevent promotion.
2. When checking for redeclarations, special-case formal parameters,
which are `var`-like but also prevent promotion.

Reviewed By: tmikov

Differential Revision: D44516271

fbshipit-source-id: 6789b30d7ed99ecdb5e07f70b17dd773686e392f