Avoid creating single type unions
Summary:
We can currently create unions with only a single element if
canonicalising the types in the union reduces to a single type. This is
a problem for arrays, which are invariant in the type of their element,
and therefore rely on equality comparison. So an array of type
`(number | number)[]` would not be able to flow into an array of type
`number[]`.

To fix this, add a new `maybeCreateUnion` and always use it when
creating unions. It will return a union if canonicalisation results in
more than one type, and otherwise just return the single resulting type.

Reviewed By: tmikov

Differential Revision: D44492593

fbshipit-source-id: f42da7ca08d3f00304e9298eaed2e99fa0cc34f3