Restructure FunctionAnalysis to use worklist
Summary:
FunctionAnalysis currently only goes one indirection deep when looking
for callsites to register. This will result in not setting all the
possible target operands correctly in some cases.

As we add casts such as UnionNarrowTrusted, we'll encounter more of
these instances, meaning we need to have FunctionAnalysis properly
iterate all the instructions whose results are known to contain the
created closure and register the callsites found along the way.

To facilitate this, write a DFS which can traverse whichever
Instructions we'd want to allow to provide "edges" in this graph,
providing "vertices" which are Instructions whose results known to
contain the created closure.

Reviewed By: neildhar

Differential Revision: D44554835

fbshipit-source-id: d3b74f935046f4aaf79ae1ae93e3cf1c5ba4b1e5