Fix node-types flow error
Summary: Within `node-types` the `asDetachedNode` function is frequently called. However with the latest version of Flow this require specific generic arguments to be passed which we don't always know. Previously this was updated to have specific Flow suppressions but the very latest Flow was erroring with a different error type. Instead of continuing to add more errors lets just create a version of this function that uses any for its inputs and outputs. This should be safe enough since the outer codegenned function is well typed so the any should not escape.

Reviewed By: SamChou19815

Differential Revision: D44770885

fbshipit-source-id: cac1b1b3b5bb6796954dabddf64a985ad2d758d0