Types require strict mode
Summary:
Make sure that typed code is always compiled in strict mode. Some of the
changes in this diff may seem redundant, but I wanted a robust solution
that wouldn't surprise us when new APIs are added.

First, using `--typed` forces strict mode.

Second, `shermes.cpp:parseJS()` temporarily overrides the strictness to
true if asked to parse a typed file. This may seem redundant, but in the
future it is conceivable that one invocation of shermes would compile
both typed and legacy modules. If the latter happen to be CommonJS, they
might be in loose mode.

Third, assert in `FlowChecker::run()` that strict mode is on and that
the global function is strict. Also to be safe even in release mode,
check it and generate an error.

Last, in FlowChecker::visitFunctionLike(), assert that the current
function is strict. Again, this may seem redundant, but
`FlowChecker::run()` may not be the only API for `FlowChecker`, so it is
better to be safe.

Reviewed By: neildhar

Differential Revision: D44565912

fbshipit-source-id: 0da24d67df6c18cf414446d71e2e32422ea777a1