Remove register stack size override in hermes.cpp
Summary:
Original Author: neildhar@meta.com
Original Git: 03f2dffc1d0ef8b2360a6790ad425ce4013e4de3

We are currently overriding any specified register stack size with a
size that is fairly low. This was added to allow the register stack to
be placed on a thread stack when StackRuntime is enabled, however, it
seems like that functionality was removed some time ago, so this limit
no longer serves a purpose.

Remove it so that sizes set through RuntimeConfig are actually
honoured. Set the default RuntimeConfig value to be roughly what the
old limit was (no need to suddenly increase the stack size if we
haven't had problems yet). Keeping the limit low helps us retain the
ability to place the register stack in StackRuntime again in the future.

Original Reviewed By: mhorowitz

Original Revision: D42610775

Reviewed By: tmikov

Differential Revision: D44771065

fbshipit-source-id: f889ae8ac5abc5e8298d684af10c4b767e619a1a