Add mayHaveAccessor field to ClassFlags
Summary:
Original Author: fbmal7@meta.com
Original Git: 79c94de63cd417e5c3d95332dd9aae95ffc6174d

Add a new field to the struct `ClassFlags`, which represents if there
is *potentially* a property accessor in the `HiddenClass`. This field
is pessimistic and can sometimes give false positives. This is due to
the fact that the field may only ever be set, but never cleared. So if
you define a getter property on an object, and then delete that
property, the field will remain set. This field will open up the door
to some potential optimizations now that we have this knowledge. It is
also used to fix a correctness bug in the next diff.

Original Reviewed By: neildhar

Original Revision: D41751159

Reviewed By: tmikov

Differential Revision: D44771061

fbshipit-source-id: 7f7bf6733a9a3255c0bd698de4f4613ba2fefe8b