fix scope resolution for react imports during ts translation
Summary: noticed this was not working in most, if not all, scenarios. `ScopeManager.acquire` matches on `scope.block`, so most nodes need more work to properly resolve. bubble upwards through the AST until we find a node with a scope. use the second argument of `ScopeManager.acquire` to specify that we're interested in the _inner_ scope, so module scope returns before global.

Reviewed By: pieterv

Differential Revision: D44720346

fbshipit-source-id: 0a834ca3938873d56d4fa91ec167fa67b1a1f9f9