Process key strings efficiently
Summary:
Original Author: fbmal7@meta.com
Original Git: f227dd62d39a931f3a39453136d58d90784379cd

If we are parsing a string that will be used as a key in an object,
then we know the string hash must be computed eventually, since it will
need to be turned into a SymbolID. However, if we are parsing the
string as a simple string value, then we can make the optimization of
side stepping the hash computation and lookup, and construct a new
StringPrimitive directly. ~7.2% perf improvement.

Original Reviewed By: tmikov

Original Revision: D43486478

Reviewed By: neildhar

Differential Revision: D44771075

fbshipit-source-id: 34da7bd1451157a7062b2c533d6a1f50088fc369