Use global capture information to refine frame load elimination
Summary:
Add a new less precise analysis that can be used on captured
loads/stores, or outside the entry block. It uses information about
whether a variable has captured loads or stores globally, to determine
whether intervening instructions that may execute JS could modify it.

The net effect is that duplicate loads in inner functions and outside
the entry block can be optimised even when there are intervening
function calls (see the updated tests).

Note that this analysis is strictly more powerful than the existing
one for "const" values. A const variable will have no captured
stores, which means that loads from it will be coalesced.

Reviewed By: tmikov

Differential Revision: D42422929

fbshipit-source-id: e90810f5ddbd956f4ba046a15e11647c83f60973