Tighten criteria for for eliminating frame loads and stores
Summary:
Instead of checking for whether an instruction can read/write memory,
we really only need to check for whether it can execute additional
code. Loads and stores to the frame are handled explicitly, which means
that the only thing we need to catch with this check is potential
captures that modify the variable. And the only way for those captures
to execute is with an instruction that may execute code.

Reviewed By: tmikov

Differential Revision: D44544618

fbshipit-source-id: bda2a04260ef4d8276be0be61c67ab7e7310154d