Erase invalidated values when optimising frame loads and stores
Summary:
Instead of writing `nullptr`, when a value is invalidated in
`knownFrameValues` or `prevStoreFrame`, delete the value entirely. This
simplifies the implementation, and should improve the scalability of
the pass on large functions, since values that are invalid no longer
stored in the map and therefore don't need to be iterated over.

Correspondingly, remove the limit on the size of these data structures.
It is still technically possible to trigger slowdowns in this pass by having
(for example) a large number of loads of variables without captured
stores, and then a large number of instructions that may execute JS.
If that does turn out to happen, we could bring back the limit.

Reviewed By: tmikov

Differential Revision: D44552227

fbshipit-source-id: fa286b98454539b267a35a6c8cddafe53443d664