Restrict native stack when testing for native overflow (#966)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/966

This test is flaky because the register stack will sometimes overflow
first. On macOS, I have observed stack sizes as large as 64MB.

I also looked at setting the stack size via a linker flag, but it turns
out that the method we use for obtaining the stack size on the main
thread on macOS does not reflect that value. This seems to be a problem
for other language implementations as well. JSC does what we're doing,
and OpenJDK has some extra logic that falls back to a hard coded value.
So it seems wise not to tamper with the linker stack size for now.

Differential Revision: D44859336

fbshipit-source-id: 7156f6f0ed437c6c2b32517c210ce76ebcf824eb